drop java-library plugins, cleanup, disable werror for java
