+++ build-logic/build-parameters/build.gradle.kts
@@ -5,32 +5,36 @@ plugins {
 }
 
 buildParameters {
-    // Other plugins can contribute parameters, so below list is not exhaustive
-    enableValidation.set(false)
     pluginId("build-logic.build-params")
-    bool("skipJavadoc") {
-        defaultValue.set(false)
-        description.set("Skip javadoc generation")
-    }
-    bool("failOnJavadocWarning") {
-        defaultValue.set(true)
-        description.set("Fail build on javadoc warnings")
-    }
-    bool("enableGradleMetadata") {
-        defaultValue.set(false)
-        description.set("Generate and publish Gradle Module Metadata")
-    }
-    bool("useGpgCmd") {
-        defaultValue.set(false)
-        description.set("By default use Java implementation to sign artifacts. When useGpgCmd=true, then gpg command line tool is used for signing artifacts")
-    }
-    bool("werror") {
-        defaultValue.set(true)
-        description.set("Treat javac, javadoc, kotlinc warnings as errors")
-    }
+
+    // Other plugins can contribute parameters, so below list is not exhaustive, hencew we disable the validation
+    enableValidation.set(false)
+
+    val defaultJdkVersion = 11
     integer("defaultJdkVersion") {
-        defaultValue.set(11)
+        defaultValue.set(defaultJdkVersion)
         mandatory.set(true)
         description.set("Default Java version for source and target compatibility")
     }
+
+    group("kotlin") {
+        bool("werror") {
+            defaultValue.set(true)
+            description.set("Treat kotlinc warnings as errors")
+        }
+    }
+
+
+    group("java") {
+        integer("version") {
+            fromEnvironment()
+            defaultValue.set(defaultJdkVersion)
+            description.set("Java version used for java.toolchain")
+        }
+        bool("werror") {
+            defaultValue.set(true)
+            description.set("Treat javac, javadoc, warnings as errors")
+        }
+    }
+
 }
