+++ build-logic/jvm/src/main/kotlin/build-logic.junit-jacoco-conventions.gradle.kts
@@ -4,6 +4,15 @@ import org.gradle.kotlin.dsl.findByType
 import org.gradle.testing.jacoco.plugins.JacocoPluginExtension
 import org.gradle.testing.jacoco.tasks.JacocoReport
 
+plugins {
+    id("ch.tutteli.gradle.plugins.junitjacoco")
+}
+
+dependencies {
+    testImplementation("org.junit.jupiter:junit-jupiter-api")
+    testImplementation("org.junit.jupiter:junit-jupiter-params")
+}
+
 val jacocoToolVersion: String by rootProject.extra
 
 plugins.withId("jacoco") {
@@ -17,6 +26,7 @@ plugins.withId("jacoco") {
     tasks.withType<JacocoReport>()
         .matching { it.name == "jacocoTestReport" }
         .configureEach {
+            //TODO 1.2.0 translate to convention-plugin (or add to tutteli-plugin directly)
             if (project.extra.has(jacocoAdditionalExtraName)) {
                 val additional = project.extra.get(jacocoAdditionalExtraName) as List<*>
                 additional.forEach { p ->
