+++ samples/multiplatform/kotlin-test/build.gradle.kts
@@ -55,7 +55,7 @@ project.tasks.withType(AbstractTestTask::class.java) {
     }
 }
 
-rootProject.plugins.withType<org.jetbrains.kotlin.gradle.targets.js.yarn.YarnPlugin> {
+rootProject.plugins.withType<org.jetbrains.kotlin.gradle.targets.js.yarn.YarnPlugin>().configureEach {
     rootProject.the<org.jetbrains.kotlin.gradle.targets.js.yarn.YarnRootExtension>().apply {
         // fix vulnerable version, use 8.1.0 instead of 7.2.0 which still relies on minimatch 3.0.4
         resolution("glob", "8.1.0")
